Migrate LEG example backend to LLVM 3.14
LEG is an example backend based on ARM developed as a tutorial for
beginners. See https://github.com/frasercrmck/llvm-leg
The project and talk are from 2014 and it no longer works with trunk
llvm. This commit is my attempt at getting the example working again

With this commit the whole thing compiles and the LEG target is
recognized by the frontend driver, but I haven't managed to get it
recognized by the backend. I need to debug the driver compile command
down to the target lookup in the backend but it seems that the driver is
a bit of a b....tch on Windows and it spawns a driver subprocess instead
of running everything directly so I can't debug it properly (That's why
I've setup a linux devcontainer, I hope to be able to debug this thing
on linux).