Fix LEG instruction lowering class
For some reason the LEG instruction lowering class was passing the MC
context as an external parameter, which was not being initialized. If
you ckeck other targets a reference to this context is available at the
asm writer so there's no need to pass the reference to the MC context
since the instlowering class already contains a ref to the asm writer.

I guess this is a change in the LLVM architecture done after the LEG
example was written.